+++ client/ts/plots/index.ts
@@ -1,4 +1,4 @@
-import { heatmap } from "./heatmap";
+import { createHeatMap } from "./heatmap";
 import { CreateRidgelineChart } from "./ridgeline";
 
 /**
@@ -19,13 +19,4 @@ export function createPlot(type:string, data:any, chartHolderId:string) {
             // TODO: Add plot function
             break;
     }
-}
-
-function createHeatMap(data:any, chartHolderId:string) {
-    let keys = Object.keys(data);
-    let xAxisTitle:string = keys[0];
-    let yAxisTitle:string = keys[1];
-    let xvals = Object.values(data[keys[0]]);
-    let yvals = Object.values(data[keys[1]]);
-    heatmap(xvals, yvals, xAxisTitle, yAxisTitle,chartHolderId);
 }
